refactor(navigation): Remove application version display from sidebar
The application version display in the navigation sidebar is no longer required and has been removed to declutter the user interface.

This commit comments out the HTML block responsible for rendering the version number in `nav-content.component.html` and clears an associated, now-unused, title property in the component's TypeScript file.